+++ src/App.js
@@ -1,10 +1,8 @@
 import React, { Component } from 'react';
 
-import './css/menu.css'
 import './css/style.css'
 import './css/normalize.css'
 import './css/flexboxgrid.css'
-
 
 import Itemimg_1 from './images/bitmap.jpg'
 import Itemimg_2 from './images/bitmap1.jpg'
@@ -26,17 +24,17 @@ class App extends Component {
               <Card />
               <div className="col-4">
                 <img className="item-image" src={Itemimg_2} alt="item-img" />
-                <a href="#" className="item-price">$240.99</a>
+                <a href="" className="item-price">$240.99</a>
               </div>
               <div className="col-4">
                 <img className="item-image" src={Itemimg_1} alt="item-img" />
-                <a href="#" className="item-price">$1024</a>
+                <a href="" className="item-price">$1024</a>
               </div>
             </div>
             <div className="row">
               <div className="col-4">
                 <img className="item-image" src={Itemimg_2} alt="item-img" />
-                <a href="#" className="item-price-sale">$170</a>
+                <a href="" className="item-price-sale">$170</a>
               </div>
               <div className="col-4">
                 <span className="sale">sale</span>
