+++ src/Sidebar/Searchbar/index.js
@@ -1,10 +1,8 @@
 import React from 'react'
 
 
-import '../css/menu.css'
-import '../css/style.css'
-import '../css/normalize.css'
-import '../css/flexboxgrid.css'
+
+import '.style.css'
 
 
 export default ()=> {
