+++ src/Sidebar/Searchbar/style.css
@@ -1,0 +1,26 @@
+.searchbar{
+  margin-top:144px;
+  background-image: url(./images/group.png);
+  background-repeat: no-repeat;
+  background-position: 14px 0px;
+  background-size: 28px 30px;
+  margin-left: 22px;
+  margin-right: 22px;
+  padding-left: 60px;
+  padding-top: 2px;
+  padding-bottom: 12px;
+  border-bottom: 4px solid #373737;
+}
+
+
+.searchbar-input-line{
+  background-color: #0e0e0e;
+  color: #fff;
+  width: 100%;
+  font-size: 24px;
+  padding:0;
+  margin:0;
+  outline:none;
+  border: none;
+  -webkit-appearance: none;
+}
