+++ src/Sidebar/index.js
@@ -3,10 +3,7 @@ import Searchbar from './Searchbar'
 import Menu from './Menu'
 
 
-import '../css/menu.css'
-import '../css/style.css'
-import '../css/normalize.css'
-import '../css/flexboxgrid.css'
+import './style.css'
 
 
 export default ()=> {
