+++ src/Sidebar/menu/style.css
@@ -1,0 +1,38 @@
+.sidebar-menu{
+  display:flex;
+  flex-direction: column;
+  align-items: center;
+  text-transform: uppercase;
+  font-size: 24px;
+  margin-top: 150px;
+}
+
+.menu-current{
+  font-family: AvenirNextBold;
+  color: #fff;
+}
+
+.menu-current img{
+  margin-left: 9px;
+  margin-bottom: 2px;
+}
+
+.submenu-current{
+  margin-top: 15px;
+  margin-bottom: 26px;
+  font-family: AndaleMono;
+  color: #fff;
+}
+
+.submenu{
+  font-family: AndaleMono;
+  color: #3c3c3c;
+  margin-bottom: 25px;
+}
+
+.menu{
+  margin-top: 25px;
+  margin-bottom: 5px;
+  font-family: AvenirNextBold;
+  color: #3c3c3c;
+}
