+++ src/Sidebar/style.css
@@ -1,0 +1,10 @@
+.sidebar{
+  margin:0;
+  padding:0;
+  width: 415px;
+  min-height: 1171px;
+  background-color: #0e0e0e;
+  background-image: url(./images/logo.png);
+  background-repeat: no-repeat;
+  background-position: center 50px;
+}
