+++ src/first page/Card/index.js
@@ -1,14 +1,10 @@
 import React from 'react'
 
-
-import './css/menu.css'
-import './css/style.css'
-import './css/normalize.css'
-import './css/flexboxgrid.css'
+import './style.css'
 
 import Itemimg from './images/bitmap1.jpg'
 
-import Sale_label from './SaleLabel'
+import Sale_label from '../SaleLabel'
 
 export default ()=> {
   return(
