+++ src/first page/Card/style.css
@@ -1,0 +1,36 @@
+.col-4{
+  margin-right: 24px;
+  padding: 9px;
+  display: flex;
+  flex-direction: column;
+  position: relative;
+  background-color: #f4f4f4;
+}
+
+.item-image{
+  height: 197px;
+  width: 294px;
+}
+
+.item-price, .item-price-sale{
+  font-size: 30px;
+  font-family: AvenirNextBold;
+  text-transform: uppercase;
+  text-align: center;
+  border: none;
+  margin-top: 6px;
+  height: 91px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.item-price{
+  background-color: #fff;
+  color: #111;
+}
+
+.item-price-sale{
+  background-image: linear-gradient(107deg, #0c09bf, #966dd8);
+  color: #fff;
+}
