--- src/header.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import logo from './logo.svg';
-
-export default ()=> {
-  return(
-    <div className="App-header">
-      <img src={logo} className="App-logo" alt="logo" />
-      <h2>Welcome to React</h2>
-    </div>
-  );
-};
